requirements.txt
# Wasabi Log Monitor - Dependencies
boto3>=1.28.0
pyyaml>=6.0
apscheduler>=3.10.0
ipaddress>=1.0.23
